+++ src/mdoc/core/Page.js
@@ -1,6 +1,5 @@
 const fs = require('fs-extra')
 const hash = require('hash-sum')
-const renderHtml = require('./renderHtml')
 const fileToPath = require('../utils/fileToPath')
 
 class Page {
@@ -20,9 +19,7 @@ class Page {
     }
 
     if (this._content) {
-      const mdHtml = markdown.render(this._content)
-      const html = await renderHtml(mdHtml, this)
-      await this._context.writeTemp(fileToPath(this._relative), html)
+      this.html = markdown.render(this._content)
     }
   }
 }